docs: update documentation with SMART implementation details
- Document platform-specific SMART support (Linux/macOS only)
- Explain build tag architecture for cross-platform compilation
- Add details about ATA string byte-swapping implementation
- Clarify why disk temperatures require elevated privileges
- Note Windows uses stub implementation for compatibility

This ensures users understand the requirements and limitations of
SMART disk temperature monitoring across different platforms.
